feat : 회원인증 db작업 다시 추가, 화이스 리스트 구현

+++ 4Week_Mission/src/main/java/com/ll/exam/final__2022_10_08/api/security/jwt/JwtProvider.java
@@ -20,7 +20,7 @@ public class JwtProvider {
     }
 
     // 토큰화
-    public String generateAccessToken(Map<String, Object> claims, int seconds) {
+    public String generateAccessToken(Map<String, Object> claims, long seconds) {
         long now = new Date().getTime();
         Date accessTokenExpiresIn = new Date(now + 1000L * seconds);
